+++ src/larcv3/app/queueio/BatchFillerSparseTensor.cxx
@@ -68,7 +68,7 @@ namespace larcv3 {
   void BatchFillerSparseTensor<dimension>::finalize() { _entry_data.clear(); }
 
   template<size_t dimension>
-  int BatchFillerSparseTensor<dimension>::_check_projection(const int & projection_id) {
+  int BatchFillerSparseTensor<dimension>::_check_projection(const size_t & projection_id) {
 
     if (dimension == 3) {
       return 0;
