+++ src/larcv3/app/queueio/BatchFillerTensor.h
@@ -55,7 +55,7 @@ namespace larcv3 {
 
     bool _process_sparse(IOManager& mgr);
     bool _process_dense(IOManager& mgr);
-    int _check_projection(const int & projection_id);
+    int _check_projection(const size_t & projection_id);
     size_t _set_image_size(const EventTensor<dimension>& image_data);
     void _assert_dimension(const EventTensor<dimension>& image_data) const;
 
@@ -109,4 +109,3 @@ namespace larcv3 {
 
 #endif
 /** @} */ // end of doxygen group
-
